fix(insights): move test helpers out of __tests__ to avoid Jest pickup
Jest treats all files in __tests__/ as test files. Move shared helpers
to hog-charts/test-helpers.ts so it's not run as a test suite.
