+++ frontend/src/lib/hog-charts/test-helpers.ts
